# Simplified main parsing function.