Factored out parts of the main parse function into separate functions.

Perf and code size appear neutral.

```
name                                      old time/op  new time/op  delta
ArenaOneAlloc                             21.0ns ± 0%  20.9ns ± 0%  -0.65%  (p=0.000 n=12+12)
ArenaInitialBlockOneAlloc                 6.28ns ± 0%  6.31ns ± 0%  +0.48%  (p=0.000 n=12+12)
LoadDescriptor_Upb                        56.8µs ± 1%  54.3µs ± 1%  -4.41%  (p=0.000 n=12+12)
LoadAdsDescriptor_Upb                     3.55ms ± 3%  3.56ms ± 3%    ~     (p=0.219 n=12+12)
LoadDescriptor_Proto2                      240µs ± 0%   239µs ± 0%  -0.24%  (p=0.000 n=12+12)
LoadAdsDescriptor_Proto2                  14.4ms ± 0%  14.4ms ± 0%  -0.34%  (p=0.000 n=11+10)
Parse_Upb_FileDesc<UseArena,Copy>         10.6µs ± 0%  10.8µs ± 0%  +2.07%  (p=0.000 n=12+11)
Parse_Upb_FileDesc<UseArena,Alias>        9.31µs ± 0%  9.27µs ± 0%  -0.43%  (p=0.000 n=12+12)
Parse_Upb_FileDesc<InitBlock,Copy>        10.2µs ± 0%  10.3µs ± 0%  +0.74%  (p=0.000 n=12+12)
Parse_Upb_FileDesc<InitBlock,Alias>       8.95µs ± 0%  8.95µs ± 0%    ~     (p=0.242 n=12+12)
Parse_Proto2<FileDesc,NoArena,Copy>       29.2µs ± 0%  29.4µs ± 0%  +0.58%  (p=0.000 n=12+12)
Parse_Proto2<FileDesc,UseArena,Copy>      20.7µs ± 3%  20.7µs ± 2%    ~     (p=0.551 n=12+12)
Parse_Proto2<FileDesc,InitBlock,Copy>     16.6µs ± 0%  16.7µs ± 0%  +0.14%  (p=0.000 n=12+12)
Parse_Proto2<FileDescSV,InitBlock,Alias>  16.7µs ± 0%  16.6µs ± 0%  -0.78%  (p=0.000 n=12+12)
SerializeDescriptor_Proto2                5.35µs ± 1%  5.36µs ± 1%    ~     (p=0.671 n=12+12)
SerializeDescriptor_Upb                   11.1µs ± 0%  11.1µs ± 0%    ~     (p=0.089 n=12+12)


    FILE SIZE        VM SIZE    
 --------------  -------------- 
  +4.5%    +128  [ = ]       0    [Unmapped]
  -1.1%    -128  -1.1%    -128    upb/decode.c
    -1.3%    -128  -1.3%    -128    decode_msg
  [ = ]       0  -0.1%    -128    TOTAL
```